Fix clearing of interrupt status
To clear bits of the interrupt status register we must specify the bits
to clear. Using the right-shifted value of the interrupt status (which
is 0), is wrong and does not clear any bits. We have to use the original
value instead.

This makes the driver work on real hardware. I tested this on an Intel
Skylake laptop and an Intel Haswell desktop machine.

Change-Id: Icd38b8060a36736838c44c07f107be98260afbd0