Trim whitespace from device serial number
Depending on how the vendor implemented the space-padding of the device
serial number there may be leading or trailing spaces. In case of leading
whitespace configuring device assignment using the serial number is
cumbersome because the user needs to fill up the serial number with
enough leading whitespace.

This change trims all whitespace from the hid string.

Fixes #723.

Change-Id: Iad5a42779e140520d2f44d750d0118d39fb2971f