Reorganzise connection list
Devices that are being set up should not appear in the
connection list, so that they are not used by incoming
clients in that half set up state. Also the application
must not use pointers into the vector while the vector
is being appended to.

Switch Connections to being Ref_objs and only add them
to the official connection list when scanning is
completed. Also keep track of how many devices are
being scanned, so that create() can return an appropriate
error if the device is not found.

As a side-effect, statically configured devices are
now given priority.

Fixes #729.
See #728.

Change-Id: I8993a1e66f8bc69cae130ea9274bd3b9eca6db6b