Align length definitions with ATA command spec
The AHCI driver defines a couple of constants for handling of the device
data info page. Reviewing these values it turned out that they are not
aligned with the ATA command specification and instead of defining a
byte length they define a word length (2 bytes).

This makes the usage of these constants error prone and in fact the
Device_info::set() function misuses the constant (see ticket #723).

This change modifies the constants to reflect their actual byte value.
The code that uses these constants is fixed to handle everything
correctly. I particular the id2str function was modified to iterate the
device info ID string correctly.

Fixes #724.

Change-Id: Ibd0ef3b71fc2a33400f3d608411e6f402a613bda