declare mode static for examples
In the examples nothing is actually built but
they rely on already built code and tests.

However, the tests are built 'shared' which would
require dependencies to stdlib-sh, which in turn
is useless as we don't build anything, so we cheat
the mode 'static' here to avoid the dependencies.

Fixes #941

Change-Id: If956fbabb8c92a859c910ba3b4270dec3d536176