Support virtio DEVICE_NEEDS_RESET status bit
The functionality of the DEVICE_NEEDS_RESET bit was previously conflated
with the FAILED status bit.

A device uses the DEVICE_NEEDS_RESET status to indicate that the device
has experienced an error from which it can't recover. A driver uses the
FAILED status to indicate that something went wrong in the guest, and it
has given up on the device.

Change-Id: Ibd2a4970bcd9a6bb953bb369f88e837706a22e07