fix AHCI port initialization on QEMU
The device signature is not available before the FIS buffer is enabled.
The Spec says: PxSIG "... contains the initial signature of an attached
device when the first D2H Register FIS is received from that device."
A FIS can only be received when a FIS buffer is set.

Fixes #CD-133.

Change-Id: Ia3ebd3732f97c2a875615a48662b1f75c0eab095