example: reorganise notification handling
This modernizes virtio-block client as follows:

* register a semaphore instead of an irq for notification
* remove code for binding and server registration
* remove server object code

The asynchronous example has been modified to no longer create
a server loop but to have its own simple read loop instead where
wait() is used when the buffers of the virtio queue are full.

As a general rule: virtio clients are supposed to use the new way
of send and wait without a server loop, also when using asynchronous
requests. If a client runs in the context of a server loop, then
it has to manually register the appropriate notification irqs and
irq handler.

Change-Id: If9b2fede0c1c1952e8a1532a49ee4872f218b46b