examples: fix test case output in release mode
The test case glue code expects a certain text to mark the test as
successfully finished. This output has to be generated using a console
output function which also works in release mode. Any 'Dbg::printf()'
output is hidden if the binary is built in release mode so use the
normal LibC function 'printf()' for the final result instead.

Fixes #CD-257.

Change-Id: I4c112c8811d13f86d40487a9d4f61335fb335ccd