Remove non-essential command list running check
The AHCI specification does not require a driver to check that the
Command List Running (PxCMD.CR) flag of a port is set before issuing a
command to the port. For example, the Linux AHCI driver does only check
the PxCMD.CR flag when disabling a port, to ensure that all command
processing is stopped (similar to Ahci_port::is_command_list_disabled()).

We, however, wait after setting the command list start (PxCMD.ST) flag
until the AHCI port sets its PxCMD.CR flag. If this does not happen
within a certain time interval, we consider it an error and disable the
port again. While this approach is equally compliant with the AHCI
specification, some implementations, e.g. the AHCI controller in the
ARM FVP simulator, do never set the PxCMD.CR flag, and thus fail to
work with our AHCI driver.

Change-Id: I6e13c5d1167dccc1e1121e91f1eb4509529fad9e