Make device_discovery more robust.
device_discovery had two failure modes: One as described in CD-272, and
one where no port for a device is a valid disk, so the device never gets
marked as scanned and the scan phase is never completed. This prevents
the "svr" cap to be bound to the server thread and no dynamic clients
can consequently be created.

This commit fixes those issues by tracking the scanning of each port
separately.

Fixes ticket #CD-272.

Change-Id: I355ba49743cfe9c0a02510c6adf213c6447178f8