increase number of maximum supported in-flight client requests
A virtio block client can issue a limited number of requests to the
AHCI driver. The current number is too low for some virtio block
clients, which then fail in the negotiation phase.

Increasing the maximum supported number of requests beyond the known
requirement of 17 to the next aligend Command_table size.

fixes ticket #CD-1126

Change-Id: Iebeb7ac1e657abbbd805a8a74251f9be63601fe8