hba: fix out-of-bounds access to HBA memory registers
The HBA memory area is not necessarily only 0x1000 bytes in size.
For 32 ports the specification states a size of 0x10ff.

The code assumed to only use 0x1000 bytes for the HBA memory. This
static assumption lead to an overlap of the last two entries in the
ports array and the Command data mapping of the first Ahci_port.

Since the HBA memory and the Command data are attached via the region
manager, the lack of allocation for HBA lead to the Command data to be
placed over the last two ports.

The HBA memory access was not checked and thus exceeded the allocation.
This whole issue did not manifest at runtime, as long as the AHCI
device did not use more than 30 ports.

For example, the QEMU AHCI device only reports a HBA memory BAR size
of 0x1000, which doesn't fit 32 ports.
Another example is the Intel Corporation Elkhart Lake SATA AHCI
controller. There the BAR size is 0x800, which fits even less ports.

Change-Id: Iafedc77071fd1e510fce893104f798a81f572889